+++ src/aks-preview/azext_aks_preview/_help.py
@@ -1719,3 +1719,192 @@ helps['aks nodepool snapshot delete'] = """
     type: command
     short-summary: Delete a nodepool snapshot.
 """
+
+helps['aks draft'] = """
+    type: group
+    short-summary: Commands to build deployment files in a project directory and deploy to an AKS cluster.
+"""
+
+helps['aks draft create'] = """
+    type: command
+    short-summary: Generate a Dockerfile and the minimum required Kubernetes deployment files (helm, kustomize, manifests) for your project directory.
+    parameters:
+        - name: --destination
+          type: string
+          short-summary: Specify the path to the project directory (default is .).
+        - name: --app
+          type: string
+          short-summary: Specify the name of the helm release.
+        - name: --language
+          type: string
+          short-summary: Specify the language used to create the Kubernetes deployment.
+        - name: --create-config
+          type: string
+          short-summary: Specify the path to the configuration file.
+        - name: --dockerfile-only
+          type: bool
+          short-summary: Only generate Dockerfile for the Kubernetes deployment.
+        - name: --deployment-only
+          type: bool
+          short-summary: Only generate deployment files (helm, kustomize, manifests) for the Kubernetes deployment.
+        - name: --path
+          type: string
+          short-summary: Automatically download and use the Draft binary at the specified location.
+    examples:
+      - name: Prompt to generate a Dockerfile and deployment files in the current directory.
+        text: az aks draft create
+      - name: Generate only the Dockerfile in the current directory.
+        text: az aks draft create --dockerfile-only=true
+      - name: Generate only the deployment files in the current directory.
+        text: az aks draft create --deployment-only=true
+      - name: Generate a Dockerfile and an deployment file in a Java project with an app name at a specific project directory.
+        text: az aks draft create --language=java --app=some_app --destination=/projects/some_project
+"""
+
+helps['aks draft setup-gh'] = """
+    type: command
+    short-summary: Set up Github OIDC for your application
+    parameters:
+        - name: --app
+          type: string
+          short-summary: Specify the Azure Active Directory applicaton name.
+        - name: --subscription-id
+          type: string
+          short-summary: Specify the Azure subscription ID.
+        - name: --resource-group
+          type: string
+          short-summary: Specify the name of the Azure resource group.
+        - name: --provider
+          type: string
+          short-summary: Specify the cloud provider (default is azure).
+        - name: --gh-repo
+          type: string
+          short-summary: Specify the the github repository (organization/repo_name).
+        - name: --path
+          type: string
+          short-summary: Automatically download and use the Draft binary at the specified location.
+    examples:
+      - name: Prompt to setup the Github OIDC for a repository.
+        text: az aks draft setup-gh
+      - name: Setup the github OIDC on Azure for a specific repository.
+        text: az aks draft setup-gh --provider=azure --gh-repo=some_organization/some_repo
+      - name: Setup the github OIDC on Azure with subscription ID and resource group.
+        text: az aks draft setup-gh --provider=azure --subscription-id=some_subscription --resource-group=some_rg
+      - name: Setup the github OIDC with an application name on Azure with subscription ID and resource group for a specific repository.
+        text: az aks draft setup-gh --app=some_app --provider=azure --subscription-id=some_subscription --resource-group=some_rg --gh-repo=some_organization/some_repo
+"""
+
+helps['aks draft generate-workflow'] = """
+    type: command
+    short-summary: Generate a Github workflow for automatic build and deploy to AKS
+    long-summary: Before running this command, Make sure you have set up Github OIDC for your application.
+                  You also need to create a resource group, a container registry and a Kubernetes cluster on Azure and
+                  link the three resources using `az aks update -n <cluster-name> -g <resource-group-name> --attach-acr <acr-name>`.
+    parameters:
+        - name: --resource-group
+          type: string
+          short-summary: Specify the name of the Azure resource group.
+        - name: --destination
+          type: string
+          short-summary: Specify the path to the project directory (default is .).
+        - name: --cluster-name
+          type: string
+          short-summary: Specify the AKS cluster name.
+        - name: --registry-name
+          type: string
+          short-summary: Specify the path to the project directory.
+        - name: --container-name
+          type: string
+          short-summary: Specify the name of the container image.
+        - name: --branch
+          type: string
+          short-summary: Specify the Github branch to automatically deploy from.
+        - name: --path
+          type: string
+          short-summary: Automatically download and use the Draft binary at the specified location.
+    examples:
+      - name: Prompt to generate a Github workflow in the current directory.
+        text: az aks draft generate-workflow
+      - name: Prompt to generate a Github workflow in a specific project directory.
+        text: az aks draft generate-workflow --destination=/projects/some_project
+      - name: Generate a Github workflow with a resource group, an AKS cluster name, a container registry name in a specific project directory.
+        text: az aks draft generate-workflow --resource-group=some_rg --cluster-name=some_cluster --registry-name=some_registry --destination=/projects/some_project
+      - name: Generate a Github workflow that deploys from the main branch with a resource group, an AKS cluster name, a container registry name, and a container image name in a specific project directory.
+        text: az aks draft generate-workflow --branch=main --resource-group=some_rg --cluster-name=some_cluster --registry-name=some_registry --container-name=some_image --destination=/projects/some_project
+"""
+
+helps['aks draft up'] = """
+    type: command
+    short-summary: Set up Github OIDC and generate a Github workflow for automatic build and deploy to AKS
+    long-summary: This command combines `az aks draft setup-gh` and `az aks draft generate-workflow`.
+                  Before running this command, create a resource group, a container registry and a Kubernetes cluster on Azure and
+                  link the three resources using `az aks update -n <cluster-name> -g <resource-group-name> --attach-acr <acr-name>`.
+    parameters:
+        - name: --app
+          type: string
+          short-summary: Specify the name of the application.
+        - name: --subscription-id
+          type: string
+          short-summary: Specify the Azure subscription ID.
+        - name: --resource-group
+          type: string
+          short-summary: Specify the name of the Azure resource group.
+        - name: --provider
+          type: string
+          short-summary: Specify the cloud provider (default is azure).
+        - name: --gh-repo
+          type: string
+          short-summary: Specify the the github repository (organization/repo_name).
+        - name: --cluster-name
+          type: string
+          short-summary: Specify the AKS cluster name.
+        - name: --registry-name
+          type: string
+          short-summary: Specify the path to the project directory.
+        - name: --container-name
+          type: string
+          short-summary: Specify the name of the container image.
+        - name: --destination
+          type: string
+          short-summary: Specify the path to the project directory (default is .).
+        - name: --branch
+          type: string
+          short-summary: Specify the Github branch to automatically deploy from.
+        - name: --path
+          type: string
+          short-summary: Automatically download and use the Draft binary at the specified location.
+    examples:
+      - name: Prompt to setup the Github OIDC then generate a Github workflow in the current directory.
+        text: az aks draft up
+      - name: Prompt to setup the Github OIDC then generate a Github workflow in a specific project directory.
+        text: az aks draft up --destination=/projects/some_project
+      - name: Prompt to setup the Github OIDC for a specific repository then generate a Github workflow in a specific project directory.
+        text: az aks draft up --gh-repo=some_organization/some_repo --destination=/projects/some_project
+"""
+
+helps['aks draft update'] = """
+    type: command
+    short-summary: Update your application to be internet accessible.
+    long-summary: This command automatically updates your yaml files as necessary so that your
+                  application will be able to receive external requests.
+    parameters:
+        - name: --host
+          type: string
+          short-summary: Specify the host of the ingress resource.
+        - name: --certificate
+          type: string
+          short-summary: Specify the URI of the Keyvault certificate to present.
+        - name: --destination
+          type: string
+          short-summary: Specify the path to the project directory (default is .).
+        - name: --path
+          type: string
+          short-summary: Automatically download and use the Draft binary at the specified location.
+    examples:
+      - name: Prompt to update the application to be internet accessible.
+        text: az aks draft update
+      - name: Prompt to update the application to be internet accessible in a specific project directory.
+        text: az aks draft update --destination=/projects/some_project
+      - name: Update the application to be internet accessible with a host of the ingress resource and a Keyvault certificate in a specific project directory.
+        text: az aks draft update --host=some_host --certificate=some_certificate --destination=/projects/some_project
+"""