+++ src/spring/azext_spring/_params.py
@@ -13,7 +13,7 @@ from ._validators import (validate_env, validate_cosmos_type, validate_resource_
                           validate_vnet, validate_vnet_required_parameters, validate_node_resource_group,
                           validate_tracing_parameters_asc_create, validate_tracing_parameters_asc_update,
                           validate_app_insights_parameters, validate_instance_count, validate_java_agent_parameters,
-                          validate_jar)
+                          validate_ingress_timeout, validate_jar)
 from ._validators_enterprise import (only_support_enterprise, validate_builder_resource, validate_builder_create,
                                      validate_builder_update, validate_build_pool_size,
                                      validate_git_uri, validate_acs_patterns, validate_config_file_patterns,
@@ -24,7 +24,8 @@ from ._validators_enterprise import (only_support_enterprise, validate_builder_r
                                      validate_build_env, validate_target_module, validate_runtime_version)
 from ._app_validator import (fulfill_deployment_param, active_deployment_exist,
                              ensure_not_active_deployment, validate_deloy_path, validate_deloyment_create_path,
-                             validate_cpu, validate_memory, fulfill_deployment_param_or_warning, active_deployment_exist_or_warning)
+                             validate_cpu, validate_build_cpu, validate_memory, validate_build_memory,
+                             fulfill_deployment_param_or_warning, active_deployment_exist_or_warning)
 from ._app_managed_identity_validator import (validate_create_app_with_user_identity_or_warning,
                                               validate_create_app_with_system_identity_or_warning,
                                               validate_app_force_set_system_identity_or_warning,
@@ -51,7 +52,9 @@ source_path_type = CLIArgumentType(nargs='?', const='.',
                                    arg_group='Source Code deploy')
 # app cpu and memory
 cpu_type = CLIArgumentType(type=str, help='CPU resource quantity. Should be 500m or number of CPU cores.', validator=validate_cpu)
-memort_type = CLIArgumentType(type=str, help='Memory resource quantity. Should be 512Mi or #Gi, e.g., 1Gi, 3Gi.', validator=validate_memory)
+memory_type = CLIArgumentType(type=str, help='Memory resource quantity. Should be 512Mi or #Gi, e.g., 1Gi, 3Gi.', validator=validate_memory)
+build_cpu_type = CLIArgumentType(type=str, help='CPU resource quantity. Should be 500m or number of CPU cores.', validator=validate_build_cpu)
+build_memory_type = CLIArgumentType(type=str, help='Memory resource quantity. Should be 512Mi or #Gi, e.g., 1Gi, 3Gi.', validator=validate_build_memory)
 
 
 # pylint: disable=too-many-statements
@@ -107,6 +110,10 @@ def load_arguments(self, _):
                    help="Create your Azure Spring Apps service in an Azure availability zone or not, "
                         "this could only be supported in several regions at the moment.",
                    default=False, is_preview=True)
+        c.argument('ingress_read_timeout',
+                   type=int,
+                   help='Ingress read timeout value in seconds. Default 300, Minimum is 1, maximum is 1800.',
+                   validator=validate_ingress_timeout)
         c.argument('build_pool_size',
                    arg_type=get_enum_type(['S1', 'S2', 'S3', 'S4', 'S5']),
                    validator=validate_build_pool_size,
@@ -148,6 +155,10 @@ def load_arguments(self, _):
 
     with self.argument_context('spring update') as c:
         c.argument('sku', arg_type=sku_type, validator=normalize_sku)
+        c.argument('ingress_read_timeout',
+                   type=int,
+                   help='Ingress read timeout value in seconds. Minimum is 1, maximum is 1800.',
+                   validator=validate_ingress_timeout)
         c.argument('app_insights_key',
                    help="Connection string (recommended) or Instrumentation key of the existing Application Insights.",
                    validator=validate_tracing_parameters_asc_update,
@@ -208,7 +219,7 @@ def load_arguments(self, _):
                    validator=validate_create_app_with_user_identity_or_warning,
                    help="Space-separated user-assigned managed identity resource IDs to assgin to an app.")
         c.argument('cpu', arg_type=cpu_type, default="1")
-        c.argument('memory', arg_type=memort_type, default="1Gi")
+        c.argument('memory', arg_type=memory_type, default="1Gi")
         c.argument('instance_count', type=int,
                    default=1, help='Number of instance.', validator=validate_instance_count)
         c.argument('persistent_storage', type=str,
@@ -333,7 +344,7 @@ def load_arguments(self, _):
 
     with self.argument_context('spring app scale') as c:
         c.argument('cpu', arg_type=cpu_type)
-        c.argument('memory', arg_type=memort_type)
+        c.argument('memory', arg_type=memory_type)
         c.argument('instance_count', type=int, help='Number of instance.', validator=validate_instance_count)
 
     for scope in ['spring app deploy', 'spring app deployment create']:
@@ -369,6 +380,10 @@ def load_arguments(self, _):
                 'container_args', help='The arguments of the container image.', nargs='*', arg_group='Custom Container')
             c.argument(
                 'build_env', build_env_type)
+            c.argument(
+                'build_cpu', arg_type=build_cpu_type, default="1")
+            c.argument(
+                'build_memory', arg_type=build_memory_type, default="2Gi")
 
     with self.argument_context('spring app deploy') as c:
         c.argument('source_path', arg_type=source_path_type, validator=validate_deloy_path)
@@ -380,7 +395,7 @@ def load_arguments(self, _):
         c.argument('skip_clone_settings', help='Create staging deployment will automatically copy settings from production deployment.',
                    action='store_true')
         c.argument('cpu', arg_type=cpu_type)
-        c.argument('memory', arg_type=memort_type)
+        c.argument('memory', arg_type=memory_type)
         c.argument('instance_count', type=int, help='Number of instance.', validator=validate_instance_count)
 
     with self.argument_context('spring app deployment') as c: