+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,20 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
-0.5.72
+0.5.74
 ++++++
 
 * Add support for web application routing.
+
+0.5.73
+++++++
+
+* Fix import issues with command group `az aks draft`
+
+0.5.72
+++++++
+
+* First public release for `az aks draft`
 
 0.5.71
 ++++++